backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.5
pydantic==2.10.3
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
aiosqlite==0.20.0
python-dateutil==2.9.0.post0
httpx==0.28.1
reportlab==4.2.5
alembic==1.14.0
pytest==8.3.4
PyJWT==2.10.1
boto3==1.35.96



